re-implement/json-reader: Implement RequestHandler.kt
