+++ src/main/kotlin/ch/fhaas/bobs_burger_api/entity/Burger.kt
@@ -1,0 +1,9 @@
+package ch.fhaas.bobs_burger_api.entity
+
+class Burger(
+    name: String,
+    wikiUrl: String,
+    val price: String,
+    val season: Season?,
+    val episode: Episode?
+): Entity(name, wikiUrl)
