+++ src/main/kotlin/ch/fhaas/bobs_burger_api/entity/Occupation.kt
@@ -1,0 +1,5 @@
+package ch.fhaas.bobs_burger_api.entity
+
+class Occupation(
+    name: String
+): Entity(name)
