+++ src/main/kotlin/ch/fhaas/bobs_burger_api/entity/PestControlTruck.kt
@@ -1,0 +1,9 @@
+package ch.fhaas.bobs_burger_api.entity
+
+class PestControlTruck(
+    name: String,
+    wikiUrl: String,
+    val season: Season?,
+    val episode: Episode?,
+    val imageUrl: String
+): Entity(name, wikiUrl)
