+++ src/main/kotlin/ch/fhaas/bobs_burger_api/entity/Relative.kt
@@ -1,0 +1,8 @@
+package ch.fhaas.bobs_burger_api.entity
+
+
+class Relative(
+    name: String,
+    val relation: String,
+    wikiUrl: String
+): Entity(name, wikiUrl)
