+++ src/main/kotlin/ch/fhaas/bobs_burger_api/entity/Season.kt
@@ -1,0 +1,5 @@
+package ch.fhaas.bobs_burger_api.entity
+
+class Season(
+    val seasonNum: Int,
+): Entity()
