+++ src/main/kotlin/ch/fhaas/bobs_burger_api/entity/VoiceActor.kt
@@ -1,0 +1,5 @@
+package ch.fhaas.bobs_burger_api.entity
+
+class VoiceActor(
+    name: String
+): Entity(name)
