--- src/main/kotlin/ch/fhaas/real_estate_rest_api/StartAction.kt
@@ -1,20 +1,0 @@
-package ch.fhaas.real_estate_rest_api
-
-import org.springframework.boot.context.event.ApplicationStartedEvent
-import org.springframework.context.ApplicationListener
-import org.springframework.stereotype.Component
-import java.util.*
-
-@Component
-class StartAction : ApplicationListener<ApplicationStartedEvent> {
-
-    override fun onApplicationEvent(event: ApplicationStartedEvent) {
-        setSystemProperties()
-    }
-
-    private fun setSystemProperties() {
-        val properties = Properties()
-        properties["java.awt.headless"] = "false"
-        System.setProperties(properties)
-    }
-}
