--- src/main/kotlin/ch/fhaas/real_estate_rest_api/entity/Burger.kt
@@ -1,9 +1,0 @@
-package ch.fhaas.real_estate_rest_api.entity
-
-class Burger(
-    name: String,
-    wikiUrl: String,
-    val price: Double,
-    val season: Season?,
-    val episode: Episode?
-): Entity(name, wikiUrl)
