--- src/main/kotlin/ch/fhaas/real_estate_rest_api/entity/Occupation.kt
@@ -1,5 +1,0 @@
-package ch.fhaas.real_estate_rest_api.entity
-
-class Occupation(
-    name: String
-): Entity(name)
