--- src/main/kotlin/ch/fhaas/real_estate_rest_api/entity/PestControlTruck.kt
@@ -1,9 +1,0 @@
-package ch.fhaas.real_estate_rest_api.entity
-
-class PestControlTruck(
-    name: String,
-    wikiUrl: String,
-    val season: Season?,
-    val episode: Episode?,
-    val imageUrl: String
-): Entity(name, wikiUrl)
