--- src/main/kotlin/ch/fhaas/real_estate_rest_api/entity/Relative.kt
@@ -1,8 +1,0 @@
-package ch.fhaas.real_estate_rest_api.entity
-
-
-class Relative(
-    name: String,
-    val relation: String,
-    wikiUrl: String
-): Entity(name, wikiUrl)
