--- src/main/kotlin/ch/fhaas/real_estate_rest_api/entity/Season.kt
@@ -1,5 +1,0 @@
-package ch.fhaas.real_estate_rest_api.entity
-
-class Season(
-    val seasonNum: Int,
-): Entity()
