--- src/main/kotlin/ch/fhaas/real_estate_rest_api/entity/StoreNextDoor.kt
@@ -1,9 +1,0 @@
-package ch.fhaas.real_estate_rest_api.entity
-
-class StoreNextDoor(
-    name: String,
-    wikiUrl: String,
-    val firstSeason: Season?,
-    val firstEpisode: Episode?,
-    val imageUrl: String
-): Entity(name, wikiUrl)
