--- src/main/kotlin/ch/fhaas/real_estate_rest_api/entity/VoiceActor.kt
@@ -1,5 +1,0 @@
-package ch.fhaas.real_estate_rest_api.entity
-
-class VoiceActor(
-    name: String
-): Entity(name)
