--- src/main/kotlin/ch/fhaas/real_estate_rest_api/services/JsonReader.kt
@@ -1,89 +1,0 @@
-package ch.fhaas.real_estate_rest_api.services
-
-import ch.fhaas.real_estate_rest_api.entity.*
-import ch.fhaas.real_estate_rest_api.services.entity_services.EpisodeService
-import kong.unirest.core.json.JSONArray
-import kong.unirest.core.json.JSONObject
-import org.springframework.stereotype.Service
-
-@Service
-class JsonReader(private val episodeService: EpisodeService) {
-
-    fun getCharacter(json: JSONObject): Character? =
-            Character(
-                name = json.getString("name"),
-                wikiUrl = json.getString("wikiUrl"),
-                relatives = getListOf(json.getJSONArray("relatives"), this::getRelative),
-                imageUrl = json.getString("url"),
-                gender = json.getString("gender"),
-                age = json.getInt("age"),
-                hair = json.getString("hair"),
-                occupation = getOccupation(json.getJSONObject("occupation")),
-                allOccupations = getListOf(json.getJSONArray("allOccupations"), this::getOccupation),
-                firstEpisode = requestEpisode(json.getInt("firstEpisode")),
-                voicedBy = getVoiceActor(json.getJSONObject("voicedBy"))
-            )
-
-    fun getVoiceActor(json: JSONObject): VoiceActor? =
-            VoiceActor(
-                name = json.getString("name"),
-            )
-
-    fun getEpisode(json: JSONObject): Episode? =
-            Episode(
-                name = json.getString("name"),
-                wikiUrl = json.getString("wikiUrl"),
-                productionCode = json.getString("productionCode"),
-                airDate = json.getString("airDate"),
-                season = getSeason(json),
-                episodeNum = json.getInt("episode"),
-                totalViewers = json.getInt("totalViewers")
-            )
-
-    fun getSeason(json: JSONObject): Season? =
-        Season(
-            seasonNum = json.getInt("season")
-        )
-
-    fun getRelative(json: JSONObject): Relative =
-            Relative(
-                name = json.getString("name"),
-                wikiUrl = json.getString("wikiUrl"),
-                relation = json.getString("relationship")
-            )
-
-    fun getOccupation(json: JSONObject): Occupation? =
-            Occupation(
-                name = json.getString("name"),
-            )
-
-    fun getBurger(json: JSONObject): Burger? =
-        Burger(
-            name = json.getString("name"),
-            wikiUrl = json.getString("wikiUrl"),
-            price = json.getDouble("price"),
-            season = getSeason(json.getJSONObject("season")),
-            episode = requestEpisode(json.getInt("episode"))
-        )
-
-    fun getPestControlTruck(json: JSONObject): PestControlTruck? =
-        PestControlTruck(
-            name = json.getString("name"),
-            wikiUrl = json.getString("wikiUrl"),
-            season = getSeason(json.getJSONObject("season")),
-            episode = requestEpisode(json.getInt("episode")),
-            imageUrl = json.getString("url")
-        )
-
-    fun <T : Entity> getListOf(json: JSONArray, getEntity: (obj: JSONObject) -> (T?)): List<T> =
-        (0 until json.length()).fold(emptyList()) { list, index ->
-            val obj = json.getJSONObject(index)
-            val entity: T? = getEntity(obj)
-            entity?.let { list + entity } ?: list
-        }
-
-    private fun requestEpisode(index: Int): Episode? {
-        val json: JSONObject = episodeService.getByIndex(index)
-        return getEpisode(json)
-    }
-}
