+++ src/test/kotlin/ch/fhaas/bobs_burger_api/MainTests.kt
@@ -1,0 +1,15 @@
+package ch.fhaas.bobs_burger_api
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+import kotlin.test.assertTrue
+
+@SpringBootTest
+class MainTests {
+
+    @Test
+    fun contextLoads() {
+        assertTrue(true)
+    }
+
+}
