+++ src/test/kotlin/ch/fhaas/bobs_burger_api/RequesterTest.kt
@@ -1,0 +1,4 @@
+package ch.fhaas.bobs_burger_api
+
+class RequesterTest {
+}
