--- src/test/kotlin/ch/fhaas/real_estate_rest_api/MainTests.kt
@@ -1,15 +1,0 @@
-package ch.fhaas.real_estate_rest_api
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-import kotlin.test.assertTrue
-
-@SpringBootTest
-class MainTests {
-
-    @Test
-    fun contextLoads() {
-        assertTrue(true)
-    }
-
-}
