--- src/test/kotlin/ch/fhaas/real_estate_rest_api/RequesterTest.kt
@@ -1,4 +1,0 @@
-package ch.fhaas.real_estate_rest_api
-
-class RequesterTest {
-}
